backend_ai/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
google-genai
